requirements.txt
openai>=1.3.8
faiss-cpu>=1.7.4
jedi>=0.19.1
numpy>=1.24.0
